Dialogue Import 3 (Final)
Finished fleshing out and integrating the basics of the dialogue system. The novel controller can be told to load in dialogue files and exit. You can track when the novel controller is speaking. Added a clock, which currently just returns the real time. Added a Day Manager, which can manage the day of the week as well as advance and start the day, loading in the correct text file. I've held off on integrating the choice/input segement as I'm not sure it'll be necessary. Integration of that system would just require the appropriate game objects in the scene.  On Play, you can click through dialogue with "Space", and once dialogue has finished, you can press "n" to advance days and start the next one.