Screen Setup
Set up the screen canvas and panels for Meeting, Mail, and Work. Also created a Desktop, which will typically be on. This also holds icons/buttons for accessing the various panels. The Screen Canvas holds the Screen Manager script, which manages these operations. I also created functions for disabling and enabling icon interaction, which ive implemented in the Dialogue Manager so that you can't interact with the computer during dialogue sequences (i.e. changing screens while in a Meeting)