Mail Window Refinements
Primarily went through and fixed the sizing and anchoring for all Mail and Window objects. General note, many UI elements will need a "stretch" style anchor, and we'll need to set the anchors to the corners of the UIobject manually to make sure it strecthes/behaves appropriately when the window is resized. I also created a "Dock" gameobject which houses all the Dock UIobjects.