Mail Configuration: Replies
Created Reply UI objects and code for allowing MailObjects to track whether they've been replied to, and to either load in reply buttons or the previously selected reply text when loading content into the Body UI objects. Also cleaned up some additional anchoring and visual issues. Mail still needs a prefab to be created and added to the mailmanager's list, from which you can load in any mail in the list with the name. I'm considering adding another method that will check all mail objects based on their tag, and load in the appropriate mail each day. In that case, the DayManager would make the call during its StartDay() method.