Dialogue Choice Configuration
Created dialogue option objects and set references. When a "choice" is indicated in the story txt file, the ChoiceScreen creates buttons using the inactive prefab in the scene. Choices currently allow for loading new chapter files, as well as modyfing the current chapter's response line. Next step should be setting up a playerData system and expanding the Dialogue System with options that allow choices to modify playerData variables