PlayerData Setup
Created a PlayerData object attached to a GamerManager. This so far just tracks Performance as an int and a Test Boolean. The dialogue system has been expanded to include commands for adjusting the performance, as well as setting the Test Bool. The performance command should work well, but the setbool one we'll have to create a new command for every bool, since I couldn't figure out how to find which bool you're referring to from the string data taken from the txt file. Cumbersome, but hopefully we won't need too many bools!