LayerController Setup
The LayerController has been created and set up in the scene, along with a Foreground object. The LayerController (when commanded via the txt file) uses its child Graphic Object to create a new image and place it as a child of the foreground. When setting the layer, you can define various aspects such as transition speed and smoothing. I've also adderd a clear layer for fading an image out, a "wait" command so that we can have the dialogue system wait and create more of a cinematic flow.