Initial MeetingWindow Setup
Still pretty early days here, but at this point we've got a simple Character prefab, which has a sprite renderer and animator component. We'll ba manually creating thew animations using the frames Oz provides, which will give us a lot of freedom about the timing. Since we're using Sprites, we'll also be able to mess with stuff later on. Anyway, currently have a "Video Feed" parent object, which will house the animated components. We'll need some additional fluff, as well as a simple button set up for starting a meeting when allowed.