Finalizing MeetingWindow Setup
Completed the Meeting Window basic functionality. We can qeueu a meeting by accessing the MeetingManager, and we can do this from dialogue scenes as well as from opening Mail items. We may need to expand access to this set up from the Work window and other scripts, but it should be a simple integration. During a meeting, a character is pulled from the character pool and activated. Animations are played through the dialogue system. The Mail objects now have a few more booleans with fields for information. Mail can start dialogue, as well as queue up a meeting. I think next up, I'll add a couple quick UI elements to show whether a meeting is queued and one for the Mail system to show how many unread you have